custom hook for sockek logic, TODO dupe message bug
